change return type increment() to double

+++ src/main/java/com/hungnd62/testmaven/TestWhile.java
@@ -20,7 +20,7 @@ public class TestWhile {
         } else return 1;
     }
 
-    static int increment(int a) {
+    static double increment(double a) {
         a = a + 1;
         return a;
     }
